Create CharacterNameController, and update all necessary routes
